# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PublicLibrary API",
    "version": "1.0.0",
    "description": "API for PublicLibrary domain"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/loan/checkoutbook": {
      "post": {
        "summary": "Execute CheckOutBook command",
        "operationId": "CheckOutBook",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckOutBookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckOutBookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/loan/{loanId}/markloanoverdue": {
      "post": {
        "summary": "Execute MarkLoanOverdue command",
        "operationId": "MarkLoanOverdue",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkLoanOverdueRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkLoanOverdueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/loan/{loanId}/returnbook": {
      "post": {
        "summary": "Execute ReturnBook command",
        "operationId": "ReturnBook",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnBookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnBookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/loan/{loanId}/returnoverduebook": {
      "post": {
        "summary": "Execute ReturnOverdueBook command",
        "operationId": "ReturnOverdueBook",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnOverdueBookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnOverdueBookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/loan/{loanId}": {
      "get": {
        "summary": "Get Loan by ID",
        "operationId": "getLoanById",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Loan found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Loan"
                }
              }
            }
          },
          "404": {
            "description": "Loan not found"
          }
        }
      }
    },
    "/hold/placehold": {
      "post": {
        "summary": "Execute PlaceHold command",
        "operationId": "PlaceHold",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceHoldRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceHoldResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/hold/{holdId}/activatehold": {
      "post": {
        "summary": "Execute ActivateHold command",
        "operationId": "ActivateHold",
        "parameters": [
          {
            "name": "holdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateHoldRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivateHoldResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/hold/{holdId}/expirehold": {
      "post": {
        "summary": "Execute ExpireHold command",
        "operationId": "ExpireHold",
        "parameters": [
          {
            "name": "holdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpireHoldRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpireHoldResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/hold/{holdId}/fulfillhold": {
      "post": {
        "summary": "Execute FulfillHold command",
        "operationId": "FulfillHold",
        "parameters": [
          {
            "name": "holdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FulfillHoldRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FulfillHoldResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/hold/{holdId}": {
      "get": {
        "summary": "Get Hold by ID",
        "operationId": "getHoldById",
        "parameters": [
          {
            "name": "holdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hold found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hold"
                }
              }
            }
          },
          "404": {
            "description": "Hold not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Loan": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "string",
            "format": "uuid"
          },
          "bookId": {
            "type": "string",
            "format": "uuid"
          },
          "patronId": {
            "type": "string",
            "format": "uuid"
          },
          "checkoutDate": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          },
          "returnDate": {
            "type": "string"
          }
        },
        "required": [
          "loanId",
          "bookId",
          "patronId",
          "checkoutDate",
          "dueDate"
        ]
      },
      "Hold": {
        "type": "object",
        "properties": {
          "holdId": {
            "type": "string",
            "format": "uuid"
          },
          "bookId": {
            "type": "string",
            "format": "uuid"
          },
          "patronId": {
            "type": "string",
            "format": "uuid"
          },
          "placedDate": {
            "type": "string"
          },
          "expirationDate": {
            "type": "string"
          },
          "queuePosition": {
            "type": "integer"
          }
        },
        "required": [
          "holdId",
          "bookId",
          "patronId",
          "placedDate",
          "expirationDate",
          "queuePosition"
        ]
      },
...
      "CheckOutBookRequest": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string",
            "format": "uuid"
          },
          "patronId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "bookId",
          "patronId"
        ]
      },
      "CheckOutBookResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "loanId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      },
      "MarkLoanOverdueRequest": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "loanId",
          "loanId"
        ]
      },
      "MarkLoanOverdueResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "loanId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      },
      "ReturnBookRequest": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "loanId",
          "loanId"
        ]
      },
      "ReturnBookResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "loanId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      },
      "ReturnOverdueBookRequest": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "loanId",
          "loanId"
        ]
      },
      "ReturnOverdueBookResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "loanId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      },
      "PlaceHoldRequest": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string",
            "format": "uuid"
          },
          "patronId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "bookId",
          "patronId"
        ]
      },
      "PlaceHoldResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "holdId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      },
      "ActivateHoldRequest": {
        "type": "object",
        "properties": {
          "holdId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "holdId",
          "holdId"
        ]
      },
      "ActivateHoldResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "holdId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      },
      "ExpireHoldRequest": {
        "type": "object",
        "properties": {
          "holdId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "holdId",
          "holdId"
        ]
      },
      "ExpireHoldResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "holdId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      },
      "FulfillHoldRequest": {
        "type": "object",
        "properties": {
          "holdId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "holdId",
          "holdId"
        ]
      },
      "FulfillHoldResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "holdId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success"
        ]
      },
...
    }
  }
}